Increase accuracy of the position bits.
* Increase their width to 48.
* Use a rolling position scheme for bits 16-48 to increase the average accuracy.

+++ code/api/index-api/src/main/java/nu/marginalia/index/client/model/results/SearchResultKeywordScore.java
@@ -31,7 +31,7 @@ public final class SearchResultKeywordScore {
     }
 
     public int positionCount() {
-        return Integer.bitCount(positions());
+        return Long.bitCount(positions());
     }
 
     public int tfIdf() {
@@ -40,7 +40,7 @@ public final class SearchResultKeywordScore {
     public int subquery() {
         return subquery;
     }
-    public int positions() {
+    public long positions() {
         return WordMetadata.decodePositions(encodedWordMetadata);
     }
