+++ code/features-convert/keyword-extraction/src/main/java/nu/marginalia/keyword/KeywordMetadata.java
@@ -56,7 +56,7 @@ class KeywordMetadata {
         if (urlKeywords.containsDomain(stemmed))
             flags.add(WordFlags.UrlDomain);
 
-        int positions = bitmask.get(stemmed);
+        long positions = bitmask.get(stemmed);
 
         return new WordMetadata(tfidf, positions, flags).encode();
     }
